reuse code for TypedValueParser impls
