+++ crates/cli/cli/src/chainspec.rs
@@ -2,12 +2,44 @@ use std::sync::Arc;
 
 use clap::builder::TypedValueParser;
 
+#[derive(Debug, Clone)]
+struct Parser<C>(std::marker::PhantomData<C>);
+
+impl<C: ChainSpecParser> TypedValueParser for Parser<C> {
+    type Value = Arc<C::ChainSpec>;
+
+    fn parse_ref(
+        &self,
+        _cmd: &clap::Command,
+        arg: Option<&clap::Arg>,
+        value: &std::ffi::OsStr,
+    ) -> Result<Self::Value, clap::Error> {
+        let val =
+            value.to_str().ok_or_else(|| clap::Error::new(clap::error::ErrorKind::InvalidUtf8))?;
+        C::parse(val).map_err(|err| {
+            let arg = arg.map(|a| a.to_string()).unwrap_or_else(|| "...".to_owned());
+            let possible_values = C::SUPPORTED_CHAINS.join(",");
+            let msg = format!(
+                "Invalid value '{val}' for {arg}: {err}.\n    [possible values: {possible_values}]"
+            );
+            clap::Error::raw(clap::error::ErrorKind::InvalidValue, msg)
+        })
+    }
+
+    fn possible_values(
+        &self,
+    ) -> Option<Box<dyn Iterator<Item = clap::builder::PossibleValue> + '_>> {
+        let values = C::SUPPORTED_CHAINS.iter().map(clap::builder::PossibleValue::new);
+        Some(Box::new(values))
+    }
+}
+
 /// Trait for parsing chain specifications.
 ///
 /// This trait extends [`clap::builder::TypedValueParser`] to provide a parser for chain
 /// specifications. Implementers of this trait must provide a list of supported chains and a
 /// function to parse a given string into a chain spec.
-pub trait ChainSpecParser: TypedValueParser<Value = Arc<Self::ChainSpec>> + Default {
+pub trait ChainSpecParser: Clone + Send + Sync + 'static {
     /// The chain specification type.
     type ChainSpec: std::fmt::Debug + Send + Sync;
 
@@ -25,4 +57,9 @@ pub trait ChainSpecParser: TypedValueParser<Value = Arc<Self::ChainSpec>> + Defa
     /// This function will return an error if the input string cannot be parsed into a valid
     /// chain spec.
     fn parse(s: &str) -> eyre::Result<Arc<Self::ChainSpec>>;
+
+    /// Produces a [`TypedValueParser`] for this chain spec parser.
+    fn parser() -> impl TypedValueParser<Value = Arc<Self::ChainSpec>> {
+        Parser(std::marker::PhantomData::<Self>)
+    }
 }
